VTX-4537: more tests
